added recursive function to track all input variables, moved codeobject to new class
