+++ reactive_kernel/kernel.py
@@ -1,7 +1,6 @@
-
 from ipykernel.kernelbase import Kernel
 import sys
-from symtable import symtable
+from .codeObject import CodeObject
 
 __version__ = '0.1.0'
 
@@ -23,12 +22,10 @@ class ReactivePythonKernel(Kernel):
 
     def do_execute(self, code, silent, store_history=True, user_expressions=None,
                    allow_stdin=False):
-        
-        symbols = symtable(code, '<string>', 'exec')
 
         if not silent:
             try: 
-                obj = Code_Object(code)
+                obj = CodeObject(code)
             except Exception as e:
                 return {
                     'status' : 'error',
@@ -36,8 +33,8 @@ class ReactivePythonKernel(Kernel):
                     'evalue' : e.expr,
                     'traceback' : sys.exc_info()
                 }
-            #stream_content = {'name': 'stdout', 'text': obj.input_vars}
-            #self.send_response(self.iopub_socket, 'stream', stream_content)
+            stream_content = {'name': 'stdout', 'text': str(obj.input_vars)}
+            self.send_response(self.iopub_socket, 'stream', stream_content)
 
         return {'status': 'ok',
                 # The base class increments the execution count
@@ -45,43 +42,3 @@ class ReactivePythonKernel(Kernel):
                 'payload': [],
                 'user_expressions': {},
                 }
-
-
-class Code_Object:
-    def __init__(self, code):
-        self.input_vars = self.input_variables(code)
-        self.output_vars = self.output_variables(code)
-        self.code = code
-
-
-    def input_variables(self, code):
-        input_vars = []
-        symbols = symtable(code, '<string>', 'exec')
-        for i in symbols.get_symbols() :
-            if(i.is_global()) :
-                input_vars.append(i)
-        return str(input_vars)
-
-    def output_variables(self, code):
-        #return one top level defined variable, only including support for one as of now 
-        output_vars = []
-        symbols = symtable(code, '<string>', 'exec')
-        for i in symbols.get_symbols() :
-            if(i.is_assigned()) :
-                output_vars.append(i)
-        if len(output_vars) == 0:
-            return
-        if len(output_vars) == 1:
-            return str(output_vars[0])
-        raise DuplicateCellAddedError()
-
-    def code(self, code):
-        return code
-
-class DuplicateCellAddedError(Exception):
-    """Dependency graph already contains this code object
-
-    Code object identity is determined by a tuple of its exported variables
-    """
-    def __init__(self):
-        self.expr = "Attempted to define more than one local variable"
